NOISSUE - Refactor downlinks service by moving helpers to separate files
